requirements.txt
fastapi
uvicorn[standard]
playwright
playwright-stealth
python-dotenv
python-telegram-bot
